django_app/requirements.txt
absl-py==0.8.1
asgiref==3.2.3
astor==0.8.1
backcall==0.1.0
bleach==1.5.0
boto==2.49.0
boto3==1.10.39
botocore==1.13.39
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.1
Django==3.0
djangorestframework==3.11.0
docutils==0.15.2
gast==0.3.2
gensim==3.8.1
grpcio==1.25.0
html5lib==0.9999999
httpie==1.0.3
idna==2.8
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.1
jmespath==0.9.4
jupyter-client==5.3.4
jupyter-core==4.6.1
Markdown==3.1.1
mecab-python-windows==0.996.3
numpy==1.17.4
parso==0.5.2
pickleshare==0.7.5
prompt-toolkit==3.0.2
protobuf==3.11.1
psycopg2==2.8.4
Pygments==2.5.2
python-dateutil==2.8.0
pytz==2019.3
pywin32==227
pyzmq==18.1.1
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.3
six==1.13.0
smart-open==1.9.0
sqlparse==0.3.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
